fix(client_utils): added close rpc command to shutdown of gui from bec_ipython_client

+++ bec_widgets/cli/client_utils.py
@@ -173,8 +173,15 @@ class BECGuiClientMixin:
 
     def close(self) -> None:
         """
-        Close the figure.
+        Close the gui window.
         """
+        if self._process is None:
+            return
+
+        self._run_rpc("close", (), wait_for_rpc_response=False)
+        while self.gui_is_alive():
+            time.sleep(0.2)
+
         self._client.shutdown()
         if self._process:
             self._process.terminate()
